+++ iDrank/BeerViewController.m
@@ -59,7 +59,7 @@
     // custom drinks. It uses a ListOfDrinks that is saved in the app that the user can edit. 
     NSMutableArray* array = [[NSMutableArray alloc] init];
     NSMutableArray* array2 = [[NSMutableArray alloc]init];
-    for(int i = 0; i < [[[navP getDrinkList]get_drink_list] count]; i++)
+    for(long i = 0; i < [[[navP getDrinkList]get_drink_list] count]; i++)
     {
         if([[[[[navP getDrinkList]get_drink_list]objectAtIndex:i]get_name] isEqualToString:@"beer"])
         {
@@ -162,7 +162,7 @@
    
     [locationMeasurements addObject:newLocation];
     
-    for(int i = 0; i < [locationMeasurements count]-1; i++)
+    for(long i = 0; i < [locationMeasurements count]-1; i++)
     {
         
         CLLocation* temp = [[self locationMeasurements] objectAtIndex:i];
